requirements.txt
asgiref==3.7.2
asyncio==3.4.3
attrs==23.1.0
boto3==1.34.4
botocore==1.34.4
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-log-request-id==2.1.0
django-redis==5.4.0
django-simple-history==3.4.0
django-storages==1.14.2
django-unfold==0.17.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Faker==20.1.0
fastavro==1.9.2
graphql-core==3.2.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
memphis-py==1.2.1
mmh3==4.0.1
nats-py==2.6.0
packaging==23.2
protobuf==4.25.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
resend==0.6.0
rpds-py==0.15.2
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.7
watchdog==3.0.0
